requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0
python-dotenv==1.0.0
aiofiles==23.2.1
crewai==0.28.0
langchain>=0.1.0
langchain-groq>=0.0.1
groq>=0.4.1
streamlit>=1.28.0

